backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.3
asyncpg==0.31.0
bcrypt==3.2.2
billiard==4.2.4
black==26.1.0
bleach==6.3.0
blinker==1.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.4
dill==0.4.1
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.4
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastapi-mail==1.4.1
fastar==0.8.0
filelock==3.20.3
flake8==7.3.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
librt==0.7.8
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
orjson==3.11.7
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pipenv==2026.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
redis==7.2.0
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.52.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
tomlkit==0.14.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
virtualenv==20.36.1
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
